Restructured for mode modularity and enabling Router usage
