+++ web/src/components/header/Header.css
@@ -1,5 +1,5 @@
 .home-banner {
-  background-image: url('../../../assets/images/banner.jpg');
+  background-image: url('../../assets/images/banner.jpg');
   background-attachment: fixed;
   background-repeat: no-repeat;
   background-size: cover;
