+++ web/src/containers/home/Home.css
@@ -1,0 +1,10 @@
+.Home {
+    width: 100%;
+    height: 100%;
+}
+
+
+.navbar-container {
+    text-align: center;
+    transform: translateY(-50%);
+}
